Refactor compatibility functions
